Application of deep neural network
We build a simple l-layer (l>2) deep neural network to classify cat vs non-cat images which shows a 80% accuracy. For 2 layer NN we get 72% accuracy. We notice that these models are simple as we are not focusing on hyperparameter issues in this case. But on a simple note, this shows why deep neural network is better than shallow neural networks.